requirements.txt
google-generativeai
matplotlib
numpy
pandas
pillow
pytesseract
tkcalendar
